### 2019-05-21 #303 @PMassicotte
Submission: gitignore
Submitting Author: Philippe Massicotte (@PMassicotte)
Repository:   https://github.com/PMassicotte/gitignore
Version submitted:   0.1.0
Version accepted: TBD

---

```
Package: gitignore
Title: Create useful .gitignore files for your project
Version: 0.1.0.9000
Authors@R: c(person("Philippe",
         "Massicotte",
         email = "pmassicotte@hotmail.com",
         role = c("aut", "cre")))
Description: Simple interface to query gitignore.io to fetch gitignore templates that can be included in the .gitignore file. More than 450 templates are currently available.
    curl,
    glue,
    crayon,
    clipr,
    clisymbols,
    here,
    xfun
    covr
```

Because it helps *facilitate use of version control*.

Anyone working with the git technology can use the package to automatically generate the `.gitignore` file.

None that I could find.

Yes. See https://github.com/ropensci/software-review/issues/299#event-2327147743
