### 2019-05-22 #303 @melvidoni
### Editor checks:

---

#### Editor comments

Hello @PMassicotte! Thanks for submitting your package. These are the editor checks performed with `goodpractice`. Please take care of them. Please, address these before I can search for reviewers, or write a justification. Tag me when you are done, please!

```
── GP gitignore ─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 84% of code lines are covered by test
    cases.

    R/gitignore_fetch_available_templates.R:17:NA
    R/gitignore_fetch_ignore_templates.R:50:NA
    R/gitignore_fetch_ignore_templates.R:51:NA
    R/gitignore_fetch_ignore_templates.R:55:NA
    R/gitignore_write_gitignore.R:10:NA
    ... and 1 more lines

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80
    characters wide. Try make your lines shorter than 80 characters

    R\gitignore_fetch_available_templates.R:13:1
    R\gitignore_fetch_ignore_templates.R:19:1
    R\gitignore_fetch_ignore_templates.R:40:1
    R\gitignore_fetch_ignore_templates.R:46:1
    R\gitignore_fetch_ignore_templates.R:51:1
    ... and 5 more lines
```

---

Reviewers:
Due date:
