### 2019-05-23 74aefd1
Fixing long lines of code
