+++ R/gitignore_fetch_available_templates.R
@@ -9,16 +9,20 @@
 #'
 #' @examples
 #' gi_fetch_available_templates()
-gi_fetch_available_templates <- function() {
-  res <- curl::curl_fetch_memory("https://www.gitignore.io/api/list?format=json")
+gi_fetch_available_templates <-
+  function(url = "https://www.gitignore.io/api/list?format=json") {
+    res <- curl::curl_fetch_memory(url)
 
-  # Something went wrong
-  if (res$status_code != 200) {
-    stop("Status code was not 200. Returned status code:", res$status_code)
+    # Something went wrong
+    if (res$status_code != 200) {
+      stop(
+        "Status code was not 200. Returned status code:",
+        res$status_code
+      )
+    }
+
+    r <- jsonlite::fromJSON(rawToChar(res$content))
+
+    templates <- tolower(sort(purrr::map_chr(r, "name")))
+    unname(templates)
   }
-
-  r <- jsonlite::fromJSON(rawToChar(res$content))
-
-  templates <- tolower(sort(purrr::map_chr(r, "name")))
-  unname(templates)
-}
