+++ R/gitignore_fetch_ignore_templates.R
@@ -16,44 +16,56 @@
 #'
 #' # The .gitignore file can be automatically modified with `append_gitignore = TRUE`
 #' gi_fetch_ignore_templates(c("R", "python", "java"))
-gi_fetch_ignore_templates <- function(template_name, copy_to_clipboard = TRUE, append_gitignore = FALSE) {
+gi_fetch_ignore_templates <-
+  function(template_name,
+             copy_to_clipboard = TRUE,
+             append_gitignore = FALSE) {
 
-  # template_name <- c("java", "sdf", "asdsdf", "R")
+    # Check if vector of char
+    stopifnot(
+      is.atomic(template_name),
+      is.character(template_name),
+      is.atomic(template_name) || is.list(template_name),
+      is.logical(copy_to_clipboard),
+      is.logical(append_gitignore)
+    )
 
-  # Check if vector of char
-  stopifnot(
-    is.atomic(template_name),
-    is.character(template_name),
-    is.atomic(template_name) || is.list(template_name),
-    is.logical(copy_to_clipboard),
-    is.logical(append_gitignore)
-  )
+    template_name <- tolower(template_name)
 
-  template_name <- tolower(template_name)
+    # Check is all template_names exist, i.e. supported by gitignore.io
+    l <- gi_fetch_available_templates()
 
-  # Check is all template_names exist, i.e. supported by gitignore.io
-  l <- gi_fetch_available_templates()
+    i <- template_name %in% l
 
-  i <- template_name %in% l
+    if (!all(i)) {
+      stop(
+        "Some template_name were not found on gitignore.io: ",
+        crayon::red$bold(paste(template_name[!i], collapse = ", "))
+      )
+    }
 
-  if (!all(i)) {
-    stop("Some template_name were not found on gitignore.io: ", crayon::red$bold(paste(template_name[!i], collapse = ", ")))
+    # Fetch the gitignore data
+    template_name <- paste(template_name, collapse = ",")
+
+    r <- curl::curl_fetch_memory(
+      glue::glue("https://www.gitignore.io/api/{template_name}")
+    )
+
+    # Copy or not into the clipboard
+    if (copy_to_clipboard && clipr::clipr_available()) {
+      clipr::write_clip(rawToChar(r$content))
+      cat(
+        crayon::green(clisymbols::symbol$bullet),
+        paste(
+          "Copied to the clipboard.",
+          "You can now paste it in your .gitignore file.\n"
+        )
+      )
+    }
+
+    if (append_gitignore) {
+      gi_write_gitignore(rawToChar(r$content))
+    }
+
+    invisible(rawToChar(r$content))
   }
-
-  # Fetch the gitignore data
-  template_name <- paste(template_name, collapse = ",")
-
-  r <- curl::curl_fetch_memory(glue::glue("https://www.gitignore.io/api/{template_name}"))
-
-  # Copy or not into the clipboard
-  if (copy_to_clipboard && clipr::clipr_available()) {
-    clipr::write_clip(rawToChar(r$content))
-    cat(crayon::green(clisymbols::symbol$bullet), "Copied to the clipboard. You can now paste it in your .gitignore file.\n")
-  }
-
-  if (append_gitignore) {
-    gi_write_gitignore(rawToChar(r$content))
-  }
-
-  invisible(rawToChar(r$content))
-}
