### 2019-05-23 #303 @PMassicotte
Thank you @melvidoni. I have fixed long lines and added more tests.
