### 2019-05-23 d231c1f
No need to parameterize the url for fetching the templates
