### 2019-05-23 #303 @aedobbyn
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

#@# Review Comments

Great idea for a package @PMassicotte and looks good to me so far! Code is succinct and readable.

A few initial thoughts:

#### Things to look into

- You've got a couple of checks that overlap with each other: [here](https://github.com/PMassicotte/gitignore/blob/74aefd188f70941513d34d042debfadf67d4dfe9/R/gitignore_fetch_ignore_templates.R#L26) and [here](https://github.com/PMassicotte/gitignore/blob/74aefd188f70941513d34d042debfadf67d4dfe9/R/gitignore_fetch_ignore_templates.R#L28). Seems to work fine if a list of template names is supplied as long as you check that each individually is of type character (the whole `template_name` will of course be of type list).

- If `copy_to_clipboard` is false, gitignore is not returned to the console as it should be (according to the docs)

- I think you need an argument for `.gitignore_file` in `gi_fetch_ignore_templates` if you want people to be able to modify a .gitignore file that doesn't live at `here::here()` (and then pass that path to the `gi_write_gitignore` line). Also on this, I might change that argument to `gitignore_file` instead of `.gitignore_file` but either way is fine.

- The test for `gi_fetch_available_templates` is currently commented out. This function is used inside `gi_fetch_ignore_templates` so it's tested every time `gi_fetch_ignore_templates` is run an isolated test for just that function would also be good to have.

#### Good things

- Tested in R GUI, also works fine (sans the `crayon` coloring, as expected)

- Love the use of `here`, `glue`, and `crayon`; makes code very readable!

#### Things to consider

- The API returns gitignore results in alphabetical order. You may want to sort these based on the order supplied by the user.

- If the user's .gitignore file doesn't exist and `append_gitignore` is true, you might consider creating a .gitignore at `here::here()` and writing to it, possibly by prompting the user at the console to make sure they want to create and write a .gitignore

- You might consider shortening the function names to just `list_available` and `fetch_template`, or something to that effect

#### Random

- For @maurolepore: just a note that you'll need latest version of `testthat` for `expect_invisible` (line 10) in test-gitignore_fetch_ignore_templates.R (earlier versions don't have that function)
