### 2019-05-24 df03284
Remove overlap tests
