### 2019-05-24 64cc458
If copy_to_clipboard is false, gitignore is now printed in the console
