### 2019-05-24 7994c79
Adding argument to gi_fetch_ignore_template so the user can speficy the filename. Also, can create the file if it does not exists.
