+++ R/gitignore_write_gitignore.R
@@ -3,22 +3,38 @@
 #' Use returned template(s) to append the existing .gitignore file.
 #'
 #' @param new_lines Template returned by `gi_fetch_ignore_templates()`.
-#' @param .gitignore_file Path of the .gitignore file to modify.
+#' @param gitignore_file Path of the .gitignore file to modify.
 gi_write_gitignore <-
   function(new_lines,
-             .gitignore_file = here::here(".gitignore")) {
-    if (!file.exists(.gitignore_file)) {
+             gitignore_file = here::here(".gitignore")) {
+
+    stopifnot(basename(gitignore_file) != ".gitignore")
+
+    if (!file.exists(gitignore_file)) {
       cat(
         crayon::red(clisymbols::symbol$bullet),
         "The .gitignore file could not be found in the project directory",
         here::here(),
+        "Would you like to create it?",
         "\n"
       )
-      stop()
+
+      response <- utils::menu(c("Yes", "No"))
+
+      if (response == 1) {
+
+        file.create(gitignore_file)
+
+      } else {
+        stop(
+         "Could not find the file: ",
+          crayon::red$bold(gitignore_file)
+        )
+      }
     }
 
     existing_lines <-
-      readLines(.gitignore_file, warn = FALSE, encoding = "UTF-8")
+      readLines(gitignore_file, warn = FALSE, encoding = "UTF-8")
 
     new_lines_splitted <- unlist(strsplit(new_lines, "\n"))
 
@@ -34,7 +50,7 @@ gi_write_gitignore <-
 
     all <- c(existing_lines, new)
 
-    xfun::write_utf8(all, .gitignore_file)
+    xfun::write_utf8(all, gitignore_file)
 
     cat(
       crayon::green(clisymbols::symbol$bullet),
