### 2019-05-24 993d182
No test at the moment for gi_fetch_available_templates. I do not have a good idea on how to test it.
