### 2019-05-24 df8b51d
Shorten function names.
