+++ man/gi_available_templates.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/gitignore_fetch_available_templates.R
-\name{gi_fetch_available_templates}
-\alias{gi_fetch_available_templates}
+\name{gi_available_templates}
+\alias{gi_available_templates}
 \title{Fetch available templates from gitignore.io}
 \usage{
-gi_fetch_available_templates()
+gi_available_templates()
 }
 \value{
 A character vector with all operating systems, IDE or programming languages supported by gitignore.io.
@@ -16,5 +16,5 @@ This return list of all operating systems, IDE or programming languages supporte
 The returned list is returned as lower case characters.
 }
 \examples{
-gi_fetch_available_templates()
+gi_available_templates()
 }
