+++ man/gi_write_gitignore.Rd
@@ -7,7 +7,7 @@
 gi_write_gitignore(new_lines, gitignore_file = here::here(".gitignore"))
 }
 \arguments{
-\item{new_lines}{Template returned by `gi_fetch_ignore_templates()`.}
+\item{new_lines}{Template returned by `gi_fetch_templates()`.}
 
 \item{gitignore_file}{Path of the .gitignore file to modify.}
 }
