+++ NAMESPACE
@@ -1,4 +1,4 @@
 # Generated by roxygen2: do not edit by hand
 
-export(gi_fetch_available_templates)
-export(gi_fetch_ignore_templates)
+export(gi_available_templates)
+export(gi_fetch_templates)

[…]
