### 2019-05-24 #303 @PMassicotte
Dear @aedobbyn I have made all the changes you suggested except the one about the order supplied by the user. I am currently just using the API as is. The other option if so make 1 call for each template (using lapply) and then combine the results. However, this will create many calls to the API instead of just one. Happy to hear your recommendations!
