### 2019-05-25 #303 @aedobbyn
Looks good!

I agree it's best to make one call to the API instead of multiple. What you could do if you want to sort the results is to do a regex search for the language in between three `###`s since those always signal the beginning of a new language template and sort based on that. You could even separate each language into an element of a list and return the list. But this is all super optional :)

Small thing I just noticed is that [this line](https://github.com/PMassicotte/gitignore/blob/df8b51dac51f19991f41df8395df7753275e49fc/R/gitignore_fetch_ignore_templates.R#L3) in the docs references an old, longer function name and doesn't include a link (`\link{}`) to that function's documentation, which I think is always useful.
