### 2019-05-26 #303 @PMassicotte
Thank you @aedobbyn. I like your idea, I will work on it. Also, good catch with the old function name! I will change it!
