### 2019-05-26 ee6512c
Work on the documentation
