+++ man/gi_available_templates.Rd
@@ -10,7 +10,7 @@ gi_available_templates()
 A character vector with all operating systems, IDE or programming languages supported by gitignore.io.
 }
 \description{
-This return list of all operating systems, IDE or programming languages supported by gitignore.io.
+This return list of all operating systems, IDE or programming languages supported by \url{https://gitignore.io/}.
 }
 \details{
 The returned list is returned as lower case characters.
