### 2019-05-28 #303 @maurolepore
@PMassicotte, thanks for submitting your great work. Here is my review. With it, I intend to help you make it even more awesome. Some of my comments may no longer be relevant, as you seem to have already incorporated @aedobbyn's comments. Please let me know what questions you have.

Thanks @melvidoni, @aedobbyn, and @noamross for following this
discussion. I value your experience. Please do let me know anything you
think could better serve @PMassicotte, you, or rOpenSci. For your
record, so far I invested about 5.5 hours.

*Please check off boxes as applicable, and elaborate in comments below.
Your review is not limited to these topics, as described in the reviewer
guide*

  - [x] As the reviewer I confirm that there are no [conflicts of
    interest](#coi) for me to review this work (If you are unsure
    whether you are in conflict, please speak to your editor *before*
    starting your review).

  - [ ] **A statement of need** clearly stating problems the software is
    designed to solve and its target audience in README
  - [x] **Installation instructions:** for the development version of
    package and any non-standard dependencies in README
  - [ ] **Vignette(s)** demonstrating major functionality that runs
    successfully locally
  - [x] **Examples** for all exported functions in R Help that run
    successfully locally
  - [ ] **Community guidelines** including contribution guidelines in
    the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports`
    and `Maintainer` (which may be autogenerated via `Authors@R`).

> ##### For packages co-submitting to JOSS
>
>   - [ ] The package has an **obvious research application** according
>     to [JOSS’s
>     definition](http://joss.theoj.org/about#submission_requirements)
>
> The package contains a `paper.md` matching [JOSS’s
> requirements](http://joss.theoj.org/about#paper_structure) with:
>
>   - [ ] **A short summary** describing the high-level functionality of
>     the software
>   - [ ] **Authors:** A list of authors with their affiliations
>   - [ ] **A statement of need** clearly stating problems the software
>     is designed to solve and its target audience.
>   - [ ] **References:** with DOIs for all those that have one
>     (e.g. papers, datasets, software).

  - [x] **Functionality:** Any functional claims of the software been
    confirmed.
  - [x] **Performance:** Any performance claims of the software been
    confirmed.
  - [ ] **Automated tests:** Unit tests cover essential functions of the
    package and a reasonable range of inputs and conditions. All tests
    pass on the local machine.
  - [ ] **Packaging guidelines**: The package conforms to the rOpenSci
    packaging guidelines

  - [ ] **The author has responded to my review and made changes to my
    satisfaction. I recommend approving this package.**

  - [ ] Should the author(s) deem it appropriate, I agree to be
    acknowledged as a package reviewer (“rev” role) in the package
    DESCRIPTION file.

#### [General comments](https://ropensci.github.io/dev_guide/reviewerguide.html#preparing-your-review)

  - Does the code comply with general principles in the Mozilla
    reviewing guide?

Mostly yes. I fail to see is whether or not this package reinvents the
wheels. It will help to document why this packages is useful and who is
the target audience.

  - Does the package comply with the rOpenSci packaging guide?

Mostly yes, but see [Review Comments](#review-comments).

  - Are there improvements that could be made to the code style?

Yes, see [Review Comments](#review-comments).

  - Is there code duplication in the package that should be reduced?

I don’t think so.

  - Are there user interface improvements that could be made?

      - In `gi_fetch_ignore_templates()`, is the argument
        `copy_to_clipboard` useful enough to justify itself?

      - In the signature of `gitignore::gi_write_gitignore()`, I see
        `.gitignore_file = here::here(".gitignore")`. This feels like a
        low level detail that might be better to hide from users. Have
        you considered using `.gitignore_file = NULL` and handling the
        details in the function’s body?

  - Are there performance improvements that could be made?

No – as far as I can tell.

  - Is the documentation (installation
    instructions/vignettes/examples/demos) clear and sufficient? Does it
    use the principle of multiple points of entry i.e. takes into
    account the fact that any piece of documentation may be the first
    encounter the user has with the package and/or the tool/data it
    wraps?

Mostly yes, but see [Review Comments](#review-comments).

  - Were functions and arguments named to work together to form a
    common, logical programming API that is easy to read, and
    autocomplete?

Yes, although the naming could be improved to better reflect intent. See
[Review Comments](#review-comments).

  - If you have your own relevant data/problem, work through it with the
    package. You may find rough edges and use-cases the author didn’t
    think about.

-----

  - **Statement of need … target audience in README**. In README I see
    *what* I can do with the gitignore package but not *why*. Neither I
    clearly see if it is for me. Can you please state explicitly the
    problem gitignore solves and the target audience?

  - **Installation instructions**. Are the instructions to install from
    CRAN useful right now?

  - **Vignette(s) … that runs successfully locally**. Please help me. I
    can’t access the vignette locally. What am I missing?

``` r
browseVignettes("gitignore")
#> No vignettes found by browseVignettes("gitignore")
```

  - **Community guidelines including contribution guidelines in the
    README or CONTRIBUTING**. In README I see a link to
    CODE\_OF\_CONDUCT.md but no details on how to contribute. Do you
    need something like `usethis::use_tidy_contributing()`

  - **Installation succeeds as documented**.
    `install.packages("gitignore")` will fail until published on CRAN.
    Do you need this right now?

``` r
install.packages("gitignore")
#> Installing package into 'C:/Users/LeporeM/Documents/R/win-library/3.6'
#> (as 'lib' is unspecified)
#> Warning: package 'gitignore' is not available (for R version 3.6.0)
```

  - **Automated tests … cover … conditions**.

Can your tests cover all `if()` statements? For example:

Can you handle error messages with dedicated functions such as `stop()`,
`rlang::abort()`, and `usethis::ui_stop()`?

``` r
# Good
expect_error(usethis::ui_stop("oh nooo!"))
#> Error in expect_error(usethis::ui_stop("oh nooo!")): could not find function "expect_error"
# Bad
expect_error(cat(crayon::red(clisymbols::symbol$bullet), "oh nooo!"))
#> Error in expect_error(cat(crayon::red(clisymbols::symbol$bullet), "oh nooo!")): could not find function "expect_error"
```

  - **The package conforms to the rOpenSci packaging guidelines**. Based
    on my understanding of the [Packaging
    guide](https://ropensci.github.io/dev_guide/building.html) I make
    some suggestions under [Details](#details).

#### Details

##### Function and argument naming

  - In `gi_fetch_available_templates()` I see

> Description
> This return list of all operating systems, IDE or programming
> languages supported by gitignore.io.

The function name does not reflect its description.
(`fetch_available_operating_systems()`, for example, reflects the
description more accurately–but of course it doesn’t capture the case
when the output refers to an IDE or programming language.)

  - In `gi_write_gitignore()` I see

> `new_lines` Template returned by ‘gi\_fetch\_ignore\_templates()’.

Why not `template` or `fetched_template`?

> `gitignore_file` Path of the .gitignore file to modify.

Why not `path`?

##### Console messages

`gi_fetch_ignore_templates("R")` Prints messages via `cat()`. Can you
use dedicated functions to handle conditions?

##### Style

  - When styling code, consider using `this style` instead of ‘this
    style’, for example, not ‘gi\_fetch\_ignore\_templates()’ but
    `gi_fetch_ignore_templates()`.)

  - Try to keep lines to a maximum of 80 characters, including in
    DESCRIPTION (see `usethis::use_tidy_description()`).

  - Can you rename files in R/ and thestthat/tests/ to match the name of
    each main function? (I see that
    R/gitignore\_fetch\_available\_templates.R and
    testthat/tests/test-gitignore\_fetch\_available\_templates.R contain
    the source and tests for the function
    `gi_fetch_available_templates()`.)

##### REAMDE

  - Can you please add links to vignettes?
  - For package development, the usethis package populates .gitignore as
    needed. Is this close enough to what the gitignore package does to
    mention usethis in README?
  - The examples show no output. Can you show what the user should
    expect to happen?

##### Documentation

  - Can you please add an example for `gi_write_gitignore()`?

  - Can you clarify `#' @return` values?

> All functions should document the **type of object returned** under
> the @return heading.
> –<https://ropensci.github.io/dev_guide/building.html#building>

In `gi_fetch_available_templates()` I see:

> Value
> A character vector with all operating systems, IDE or programming
> languages supported by gitignore.io.

But the returned value is of type character. Referring to it as a list
may be confusing.

``` r
# Returns a character string
typeof(gi_fetch_available_templates())
#> [1] "character"
```

  - Can you please add top-level documentation?

> The package should contain top-level documentation for ?foobar, (or
> ?`foobar-package` if there is a naming conflict) …
> `usethis::use_package_doc()` adds the template for the top-level
> documentation.
> –<https://ropensci.github.io/dev_guide/building.html#building>

  - The vignettes does explain how to use the package but not why. It
    lacks motivation. What is the use case?

> The package should contain at least one vignette (…) illustrating
> realistic use cases.
> –<https://ropensci.github.io/dev_guide/building.html#building>

> If your package connects to a … online service … it should provide
> enough information for users to understand the nature of the …
> service. (…). Any vignette should outline prerequisite knowledge to be
> able to understand the vignette upfront.
> –<https://ropensci.github.io/dev_guide/building.html#building>

##### Documentation website

  - Will buildignore have a website? (see
    `?pkgdown::deploy_site_github()`).

##### Package dependencies

The number of dependencies seems high relative to the number of
features. Have you considered the trade-offs explained
[here](https://ropensci.github.io/dev_guide/building.html#package-dependencies)?

##### Miscellaneous CRAN gotchas

In DESCRIPTION:

  - Please ensure `Title:` uses title case.
  - Please enclose links in `<angle brackets>`.
  - Please ensure lines are 80 characters or shorter.
