### 2019-05-28 82e6c63
Remove CRAN installation instruction
