+++ README.Rmd
@@ -30,13 +30,7 @@ For any project, it is therefore important to have a `.gitignore` file that is c
 
 ## Installation
 
-You can install the released version of gitignore from [CRAN](https://CRAN.R-project.org) with:
-
-```{r, eval=FALSE}
-install.packages("gitignore") # Not currently published on CRAN
-```
-
-Or the dev version with: 
+`gitignore` can be installed from GitHub:
 
 ```{r, eval=FALSE}
 install.packages("devtools")
