+++ R/gitignore_fetch_available_templates.R
@@ -1,10 +1,10 @@
 #' Fetch available templates from gitignore.io
 #'
-#' This return list of all operating systems, IDE or programming languages supported by gitignore.io.
+#' This return list of all templates supported by gitignore.io.
 #'
 #' @details The returned list is returned as lower case characters.
 #'
-#' @return A character vector with all operating systems, IDE or programming languages supported by gitignore.io.
+#' @return A character vector with all templates supported by gitignore.io.
 #' @export
 #'
 #' @examples
