+++ man/gi_available_templates.Rd
@@ -7,10 +7,10 @@
 gi_available_templates()
 }
 \value{
-A character vector with all operating systems, IDE or programming languages supported by gitignore.io.
+A character vector with all templates supported by gitignore.io.
 }
 \description{
-This return list of all operating systems, IDE or programming languages supported by gitignore.io.
+This return list of all templates supported by gitignore.io.
 }
 \details{
 The returned list is returned as lower case characters.
