+++ README.Rmd
@@ -62,6 +62,13 @@ Multiple templates can be fetched by specifying multiple values:
 gi_fetch_templates(c("java", "c++"))
 ```
 
+More examples are provided in the vignette.
+
+```{r, eval=FALSE}
+browseVignettes("gitignore")
+```
+
+
 ## Code of conduct
 
 Please note that the 'gitignore' project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing to this project, you agree to abide by its terms.
