### 2019-05-28 93b0414
Add example for gi_write_gitignore()
