+++ R/gitignore_write_gitignore.R
@@ -4,6 +4,11 @@
 #'
 #' @param fetched_template Template(s) returned by `gi_fetch_templates()`.
 #' @param gitignore_file Path of the .gitignore file to modify.
+#'
+#' @examples
+#' f <- tempfile(pattern = "", fileext = ".gitignore")
+#' new_lines <- gi_fetch_templates("r")
+#' gi_write_gitignore(new_lines, f)
 gi_write_gitignore <-
   function(fetched_template,
              gitignore_file = here::here(".gitignore")) {
