+++ man/gi_write_gitignore.Rd
@@ -15,3 +15,8 @@ gi_write_gitignore(fetched_template,
 \description{
 Use the returned template(s) to append the existing .gitignore file.
 }
+\examples{
+f <- tempfile(pattern = "", fileext = ".gitignore")
+new_lines <- gi_fetch_templates("r")
+gi_write_gitignore(new_lines, f)
+}
