+++ man/gitignore-package.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/gitignore-package.R
+\docType{package}
+\name{gitignore-package}
+\alias{gitignore}
+\alias{gitignore-package}
+\title{gitignore: Create useful .gitignore files for your project}
+\description{
+Simple interface to query gitignore.io to fetch
+    gitignore templates that can be included in the .gitignore file. More
+    than 450 templates are currently available.
+}
+\seealso{
+Useful links:
+\itemize{
+  \item \url{https://github.com/PMassicotte/gitignore}
+  \item Report bugs at \url{https://github.com/PMassicotte/gitignore/issues}
+}
+
+}
+\author{
+\strong{Maintainer}: Philippe Massicotte \email{pmassicotte@hotmail.com}
+
+}
+\keyword{internal}
