### 2019-05-28 7c95390
Do not run example in gi_write_gitignore.
