+++ man/gi_write_gitignore.Rd
@@ -16,7 +16,9 @@ gi_write_gitignore(fetched_template,
 Use the returned template(s) to append the existing .gitignore file.
 }
 \examples{
+\dontrun{
 f <- tempfile(pattern = "", fileext = ".gitignore")
 new_lines <- gi_fetch_templates("r")
 gi_write_gitignore(new_lines, f)
+}
 }
