### 2019-05-28 e0350d1
se message() and warning() to communicate with the user in your functions. Please do not use print() or cat() unless it’s for a print.*() method, as these methods of printing messages are harder for the user to suppress.
