### 2019-05-28 #303 @PMassicotte
Thank you @maurolepore I started to address your comments.
