### 2019-05-30 #303 @PMassicotte
@maurolepore @aedobbyn Can I add you are "rev" in the DESCRIPTION file?
