+++ .travis.yml
@@ -2,3 +2,9 @@
 
 language: R
 cache: packages
+
+before_deploy: Rscript -e 'remotes::install_cran("pkgdown")'
+deploy:
+  provider: script
+  script: Rscript -e 'pkgdown::deploy_site_github()'
+  skip_cleanup: true
