+++ pkgdown.yml
@@ -1,0 +1,6 @@
+pandoc: '2.2'
+pkgdown: 1.3.0
+pkgdown_sha: ~
+articles:
+  introduction: introduction.html
+
