### 2019-05-31 #303 @maurolepore
Sure

On Thu, 30 May 2019 at 3:50 pm Philippe Massicotte <notifications@github.com>
wrote:

> @maurolepore <https://github.com/maurolepore> @aedobbyn
> <https://github.com/aedobbyn> Can I add you are "rev" in the DESCRIPTION
> file?
>
