### 2019-05-31 bd7e7e4
Adding Amanda and Mauro as package reviewer
