+++ man/gitignore-package.Rd
@@ -21,5 +21,11 @@ Useful links:
 \author{
 \strong{Maintainer}: Philippe Massicotte \email{pmassicotte@hotmail.com}
 
+Other contributors:
+\itemize{
+  \item Amanda Dobbyn \email{amanda.e.dobbyn@gmail.com} [reviewer]
+  \item Mauro Lepore \email{maurolepore@gmail.com} [reviewer]
+}
+
 }
 \keyword{internal}
