### 2019-05-31 351dff8
Calrify 'return' of the gi_write_gitignore function
