+++ man/gi_write_gitignore.Rd
@@ -12,6 +12,9 @@ gi_write_gitignore(fetched_template,
 
 \item{gitignore_file}{Path of the .gitignore file to modify.}
 }
+\value{
+TRUE if successed to write/append the .gitignore, FALSE otherwise.
+}
 \description{
 Use the returned template(s) to append the existing .gitignore file.
 }
