### 2019-05-31 8dde0b4
Add more justifications on why gitignore is needed.
