### 2019-05-31 #303 @PMassicotte
Hi @melvidoni! I just finished my corrections based on @maurolepore suggestions. The only thing I was not able to address is the test to cover the line he highlighted. This is because, based on the first suggestion of @maurolepore, the package is now asking the user if he/she wants to create the missing `.gitignore` file. This is done with the `menu()` for which I could not find a solution to test it automatically. I am open to any suggestions!
