### 2019-05-31 #303 @aedobbyn
Since it's interactive and tough to test in a non-interactive way, I think you'd be fine [wrapping it in `nocov` comments](https://github.com/r-lib/covr#exclusion-comments). If anyone has an idea about how to construct a test for it, of course that is even better.
