### 2019-06-03 #303 @maurolepore
@melvidoni and @PMassicotte,

* [  ] The author has responded to my review and made changes to my satisfaction. I recommend approving this package.

--

Although I don't need to see it again, I still have the following recommendations:

* Provide the website's link (http://www.pmassicotte.com/gitignore/) in the GitHub repo.
* Link .github/CONTRIBUTING.md from README.
* Reflow  documentation lines to 80 characters or less.
* Rename files to reflect the name of the main function they contain.
