### 2019-06-03 #303 @maurolepore
@PMassicotte, BTW, if you will submit to CRAN I also recommend you to ASAP go through the steps proposed by `devtools::release()` or `usethis::use_release_issue()`. The earlier you pick up those little CRAN gotchas the better.
