### 2019-06-03 c2e9c49
Add link to CONTRIBUTING
