+++ README.md
@@ -77,123 +77,14 @@ Templates can be fetched using the `gi_fetch_templates()` function.
 
 ``` r
 gi_fetch_templates("R")
-
-# Created by https://www.gitignore.io/api/r
-# Edit at https://www.gitignore.io/?templates=r
-
-### R ###
-# History files
-.Rhistory
-.Rapp.history
-
-# Session Data files
-.RData
-
-# User-specific files
-.Ruserdata
-
-# Example code in package build process
-*-Ex.R
-
-# Output files from R CMD build
-/*.tar.gz
-
-# Output files from R CMD check
-/*.Rcheck/
-
-# RStudio files
-.Rproj.user/
-
-# produced vignettes
-vignettes/*.html
-vignettes/*.pdf
-
-# OAuth2 token, see https://github.com/hadley/httr/releases/tag/v0.3
-.httr-oauth
-
-# knitr and R markdown default cache directories
-/*_cache/
-/cache/
-
-# Temporary files created by R markdown
-*.utf8.md
-*.knit.md
-
-### R.Bookdown Stack ###
-# R package: bookdown caching files
-/*_files/
-
-# End of https://www.gitignore.io/api/r
+● Copied to the clipboard. You can now paste it in your .gitignore file.
 ```
 
 Multiple templates can be fetched by specifying multiple values:
 
 ``` r
 gi_fetch_templates(c("java", "c++"))
-
-# Created by https://www.gitignore.io/api/java,c++
-# Edit at https://www.gitignore.io/?templates=java,c++
-
-### C++ ###
-# Prerequisites
-*.d
-
-# Compiled Object files
-*.slo
-*.lo
-*.o
-*.obj
-
-# Precompiled Headers
-*.gch
-*.pch
-
-# Compiled Dynamic libraries
-*.so
-*.dylib
-*.dll
-
-# Fortran module files
-*.mod
-*.smod
-
-# Compiled Static libraries
-*.lai
-*.la
-*.a
-*.lib
-
-# Executables
-*.exe
-*.out
-*.app
-
-### Java ###
-# Compiled class file
-*.class
-
-# Log file
-*.log
-
-# BlueJ files
-*.ctxt
-
-# Mobile Tools for Java (J2ME)
-.mtj.tmp/
-
-# Package Files #
-*.jar
-*.war
-*.nar
-*.ear
-*.zip
-*.tar.gz
-*.rar
-
-# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
-hs_err_pid*
-
-# End of https://www.gitignore.io/api/java,c++
+● Copied to the clipboard. You can now paste it in your .gitignore file.
 ```
 
 More examples are provided in the vignette.
@@ -208,5 +99,5 @@ website](http://www.pmassicotte.com/gitignore/).
 ## Code of conduct
 
 Please note that the ‘gitignore’ project is released with a [Contributor
-Code of Conduct](CODE_OF_CONDUCT.md). By contributing to this project,
-you agree to abide by its terms.
+Code of Conduct](CODE_OF_CONDUCT.md). By [contributing to this
+project](.github/CONTRIBUTING.md), you agree to abide by its terms.
