+++ README.Rmd
@@ -72,4 +72,4 @@ You can also visit the [gitignore website](http://www.pmassicotte.com/gitignore/
 
 ## Code of conduct
 
-Please note that the 'gitignore' project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing to this project, you agree to abide by its terms.
+Please note that the 'gitignore' project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By [contributing to this project](.github/CONTRIBUTING.md), you agree to abide by its terms.
