+++ man/gi_fetch_templates.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gitignore_fetch_ignore_templates.R
+% Please edit documentation in R/gi_fetch_templates.R
 \name{gi_fetch_templates}
 \alias{gi_fetch_templates}
 \title{Fetch gitignore template(s) from gitignore.io}
