+++ man/gi_write_gitignore.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gitignore_write_gitignore.R
+% Please edit documentation in R/gi_write_gitignore.R
 \name{gi_write_gitignore}
 \alias{gi_write_gitignore}
 \title{Append or create a .gitignore file}
