+++ man/gi_available_templates.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gitignore_fetch_available_templates.R
+% Please edit documentation in R/gi_available_templates.R
 \name{gi_available_templates}
 \alias{gi_available_templates}
 \title{Fetch available templates from gitignore.io}
