### 2019-06-14 #303 @PMassicotte
Hi everyone. Just a quick message to see if I need to do anything else at this point? This is my first submission to ropensci so I am not sure :)
