### 2019-06-14 #303 @maurolepore
@PMassicotte, good on you for following this up. I don't know, as this is also the first time I review an rOpenSci package. I look forward to seeing what the next stem is.
