### 2019-06-17 #303 @aedobbyn
Things are looking good @PMassicotte! Question -- is there a reason you're no longer exporting `gi_write_gitignore`?
