### 2019-06-17 7c9a71e
Re-export git_write_gitignore
