### 2019-06-17 #303 @aedobbyn
Don't you need

`stopifnot(basename(gitignore_file) == ".gitignore")` instead of `!=`

[here](https://github.com/PMassicotte/gitignore/blob/7c9a71e0d47a71c793f2e015b9ecb7419341720a/R/gi_write_gitignore.R#L22)? This seems to error when the basename of the file *is* `.gitignore`.
