### 2019-06-18 e84a3b0
Correctly veryfing if the file if a .gitignore file
