+++ tests/testthat/test_gi_fetch_templates.R
@@ -32,7 +32,8 @@ test_that("Template can be copied in the clipboard", {
 })
 
 test_that("A non existing .gitignore file can be created", {
-  f <- tempfile(pattern = "", fileext = ".gitignore")
+  f <- paste0(tempdir(), "/.gitignore")
+  unlist(f)
   file.create(f)
 
   expect_invisible(gi_fetch_templates("R", copy_to_clipboard = FALSE, append_gitignore = TRUE, gitignore_file = f))
