### 2019-06-18 #303 @aedobbyn
Nice, looks good! I think once you include some examples in your readme and pkgdown this will be good to go.
