### 2019-06-18 #303 @PMassicotte
Hi @aedobbyn You would like more examples?

https://github.com/PMassicotte/gitignore
