### 2019-06-18 #303 @aedobbyn
Sorry I should have been more specific -- more examples of `gi_write_gitignore` now that's it's been exported.
