### 2019-06-18 bb0e14e
Add more examples
