+++ R/gi_write_gitignore.R
@@ -11,7 +11,7 @@
 #'
 #' @examples
 #' \dontrun{
-#' f <- tempfile(pattern = "", fileext = ".gitignore")
+#' f <- paste0(tempdir(), "/.gitignore")
 #' new_lines <- gi_fetch_templates("r")
 #' gi_write_gitignore(new_lines, f)
 #' }
