+++ man/gi_write_gitignore.Rd
@@ -20,7 +20,7 @@ Use the returned template(s) to append the existing .gitignore file.
 }
 \examples{
 \dontrun{
-f <- tempfile(pattern = "", fileext = ".gitignore")
+f <- paste0(tempdir(), "/.gitignore")
 new_lines <- gi_fetch_templates("r")
 gi_write_gitignore(new_lines, f)
 }
