### 2019-06-20 b4414c0
Add ropensci badge.
