+++ README.md
@@ -5,6 +5,7 @@
 
 <!-- badges: start -->
 
+[![ropensci\_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
 [![CRAN
 status](https://www.r-pkg.org/badges/version/gitignore)](https://cran.r-project.org/package=gitignore)
 [![License: GPL
@@ -62,9 +63,9 @@ Show the first 25 templates returned by `gi_available_templates()`.
 library(gitignore)
 
 head(gi_available_templates(), 25)
-#>  [1] "1c"                   "1c-bitrix"            "actionscript"        
-#>  [4] "ada"                  "adobe"                "advancedinstaller"   
-#>  [7] "a-frame"              "agda"                 "al"                  
+#>  [1] "1c"                   "1c-bitrix"            "a-frame"             
+#>  [4] "actionscript"         "ada"                  "adobe"               
+#>  [7] "advancedinstaller"    "agda"                 "al"                  
 #> [10] "alteraquartusii"      "altium"               "android"             
 #> [13] "androidstudio"        "angular"              "anjuta"              
 #> [16] "ansible"              "apachecordova"        "apachehadoop"        
@@ -77,14 +78,123 @@ Templates can be fetched using the `gi_fetch_templates()` function.
 
 ``` r
 gi_fetch_templates("R")
-● Copied to the clipboard. You can now paste it in your .gitignore file.
+
+# Created by https://www.gitignore.io/api/r
+# Edit at https://www.gitignore.io/?templates=r
+
+### R ###
+# History files
+.Rhistory
+.Rapp.history
+
+# Session Data files
+.RData
+
+# User-specific files
+.Ruserdata
+
+# Example code in package build process
+*-Ex.R
+
+# Output files from R CMD build
+/*.tar.gz
+
+# Output files from R CMD check
+/*.Rcheck/
+
+# RStudio files
+.Rproj.user/
+
+# produced vignettes
+vignettes/*.html
+vignettes/*.pdf
+
+# OAuth2 token, see https://github.com/hadley/httr/releases/tag/v0.3
+.httr-oauth
+
+# knitr and R markdown default cache directories
+/*_cache/
+/cache/
+
+# Temporary files created by R markdown
+*.utf8.md
+*.knit.md
+
+### R.Bookdown Stack ###
+# R package: bookdown caching files
+/*_files/
+
+# End of https://www.gitignore.io/api/r
 ```
 
 Multiple templates can be fetched by specifying multiple values:
 
 ``` r
 gi_fetch_templates(c("java", "c++"))
-● Copied to the clipboard. You can now paste it in your .gitignore file.
+
+# Created by https://www.gitignore.io/api/java,c++
+# Edit at https://www.gitignore.io/?templates=java,c++
+
+### C++ ###
+# Prerequisites
+*.d
+
+# Compiled Object files
+*.slo
+*.lo
+*.o
+*.obj
+
+# Precompiled Headers
+*.gch
+*.pch
+
+# Compiled Dynamic libraries
+*.so
+*.dylib
+*.dll
+
+# Fortran module files
+*.mod
+*.smod
+
+# Compiled Static libraries
+*.lai
+*.la
+*.a
+*.lib
+
+# Executables
+*.exe
+*.out
+*.app
+
+### Java ###
+# Compiled class file
+*.class
+
+# Log file
+*.log
+
+# BlueJ files
+*.ctxt
+
+# Mobile Tools for Java (J2ME)
+.mtj.tmp/
+
+# Package Files #
+*.jar
+*.war
+*.nar
+*.ear
+*.zip
+*.tar.gz
+*.rar
+
[…]
