### 2019-06-20 #303 @PMassicotte
I transferred it. I am trying to find how to make codecov and appveyor work now :)
