### 2019-06-20 4ce3d27
Place ropensci badge at the bottom.
