### 2019-06-20 #303 @melvidoni
@PMassicotte Just added you to the team. Please, tackle all the To-Dos pointed out on my previous comment.
