### 2019-06-20 e489b94
Fixing badges
