### 2019-06-20 #303 @PMassicotte
Badges have been added and a release has been done (with Zenodo badge).
