### 2019-06-21 #303 @maurolepore
@PMassicotte, next week I'll be volunteering most of my time to work on open source projects. If you need a hand in this project let me know -- maybe it's time to work towards a CRAN submission?
