### 2019-06-21 #303 @maurolepore
@PMassicotte,
If you mean for the submission to CRAN, then I think a good way to know what might be missing is to walk throught the list produced via `usethis::use_release_issue()` and  `devtools::release()`.

I would also recommend to configure .travis.yml to check on multiple versions of R.

I could help with this next week.
