### 2019-06-21 #303 @maurolepore
BTW, this address resonds with error 404: https://www.pmassicotte.com/gitignore/

I guess it will eventually be https://docs.ropensci.org/gitignore/?

![image](https://user-images.githubusercontent.com/5856545/59929581-bb9f3a80-9406-11e9-8951-a85420a4b1ea.png)
