### 2019-06-25 #303 @stefaniebutland
@PMassicotte Your docs page is up now https://docs.ropensci.org/gitignore/ :-)
