### 2019-06-25 cbb5d0a
(TO REVERT) Add temporary files to guide checks:
* Add extended check list.
* Add jobs.R
* buildignore jobs.R and tmp.R.
* Change maintainer email.
