### 2019-06-25 #303 @PMassicotte
@stefaniebutland I have updated the URL on the github page and also in the README. I also provided instruction to install it from ropensci `devtools::install_github("ropensci/gitignore")`.
