+++ R/gi_write_gitignore.R
@@ -5,7 +5,7 @@
 #' @param fetched_template Template(s) returned by `gi_fetch_templates()`.
 #' @param gitignore_file Path of the .gitignore file to modify.
 #'
-#' @return TRUE if successed to write/append the .gitignore, FALSE otherwise.
+#' @return TRUE if succeeds to write/append the .gitignore, FALSE otherwise.
 #'
 #' @export
 #'
