### 2019-06-26 6d7d8cb
Revert "(TO REVERT) Add temporary files to guide checks:"
This reverts commit cbb5d0a8504869ba2eb64bd993e291b470522b0c.
