### 2019-06-26 a00f999
Fixing invalid file URIs
