### 2019-06-26 b0ccbf6
Fixing invalid file URIs
