+++ NEWS.md
@@ -1,3 +1,5 @@
+# gitignore 0.1.2
+
 # gitignore 0.1.1.9000
 
 * First release on CRAN.
