+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: gitignore
 Title: Create Useful .gitignore Files for your Project
-Version: 0.1.1.9000
+Version: 0.1.2
 Authors@R: 
     c(person(given = "Philippe",
              family = "Massicotte",
