### 2019-06-28 da368e9
Do not write to the clipboard by default
