### 2019-06-28 2651d18
Update installation instruction
