+++ NEWS.md
@@ -2,8 +2,6 @@
 
 # gitignore 0.1.2
 
-# gitignore 0.1.1.9000
-
 * First release on CRAN.
 
 # gitignore 0.1.1
