### 2019-06-28 #303 @PMassicotte
Accepted on CRAN!
