### 2019-06-28 9cb7f3c
URLs now point to ropensci/gitignore
