### 2019-07-17 547aec6
Trying to remove the use of the clipboard on linux
