### 2019-07-17 ef314c6
Simplify the matrix in Travis.
