+++ cran-comments.md
@@ -1,7 +1,7 @@
 ## Test environments
 
-* local OS X install, R 3.6.0
-* ubuntu 14.04 (on travis-ci), R 3.2, 3.3, 3.4, oldrel, release, devel
+* local OS X install, R 3.6.1
+* ubuntu 14.04 (on travis-ci), oldrel, release, devel
 * win-builder (devel and release)
 
 ## R CMD check results
@@ -16,7 +16,4 @@ There are currently no reverse dependencies for this package.
 
 ## Re-submission
 
-This is a re-submission fixing two minor elements pointed out by Swetlana Herbrandt.
-
-- Please replace \dontrun{} by \donttest{}
-- Please do not write to the clipboard by default (copy_to_clipboard = FALSE).
+The package was archived because it was using the clipboard on Linux. This new release fixes this.
