+++ .travis.yml
@@ -40,10 +40,6 @@ matrix:
         provider: script
         script: Rscript -e 'pkgdown::deploy_site_github()'
         skip_cleanup: true
-  - r: oldrel
-  - r: 3.4
-  - r: 3.3
-  - r: 3.2
 
 before_script:
   - sh -e /etc/init.d/xvfb start
