### 2019-07-18 358856c
More fixes to prevent the use of the clipboard on Linux and also remove temperary created files in examples
