+++ cran-comments.md
@@ -1,7 +1,7 @@
 ## Test environments
 
 * local OS X install, R 3.6.1
-* ubuntu 14.04 (on travis-ci), oldrel, release, devel
+* Ubuntu 14.04 (on Travis-ci), oldrel, release, devel
 * win-builder (devel and release)
 
 ## R CMD check results
@@ -16,4 +16,9 @@ There are currently no reverse dependencies for this package.
 
 ## Re-submission
 
-The package was archived because it was using the clipboard on Linux. This new release fixes this.
+The package was archived because it was using the clipboard on Linux. This new release fixes this. Explicitly:
+
+- Tests using the clipboard are now skipped on Linux.
+- The code in the vignette do not use the clipboard.
+- The code examples do not use the clipboard.
+- Temporary files created for the examples are now deleted using the `unlink()` function.
