### 2019-07-18 26a67a0
Update cran-comments.md.
