+++ DESCRIPTION
@@ -6,7 +6,8 @@ Authors@R:
     c(person(given = "Philippe",
              family = "Massicotte",
              role = c("aut", "cre"),
-             email = "pmassicotte@hotmail.com"),
+             email = "pmassicotte@hotmail.com",
+             comment = c(ORCID = "0000-0002-5919-4116")),
       person(given = "Amanda",
              family = "Dobbyn",
              role = "rev",
