### 2019-07-19 ad31bc0
Fixes wrong URL for codecov badge.
