### 2019-07-19 dc94376
Update NEWS.
