+++ NEWS.md
@@ -1,4 +1,6 @@
-# gitignore (development version)
+# gitignore 0.1.3
+
+* This is a minor update that prevent the use of the clipboard on CRAN Linux systems.
 
 # gitignore 0.1.2
 
