+++ NEWS.md
@@ -1,3 +1,5 @@
+# gitignore 0.1.4
+
 # gitignore 0.1.3
 
 * This is a minor update that prevent the use of the clipboard on CRAN Linux systems.
